+++ src/earthkit/geo/grids/utils/memcache.py
