+++ tests/regrid/inventory/test_gridspec.py
@@ -8,7 +8,7 @@
 
 import pytest
 
-from earthkit.geo.regrid.backends.db import SYS_DB
+from earthkit.geo.grids._regrid.backends.db import SYS_DB
 
 
 @pytest.mark.parametrize(
