+++ tests/regrid/regrid_mir/test_regrid_fieldlist.py
@@ -10,8 +10,8 @@
 import numpy as np
 import pytest
 
+from earthkit.geo import regrid
 from earthkit.geo.grids import Grid
-from earthkit.geo.regrid import regrid
 from earthkit.geo.utils.testing import (
     NO_EKD,  # noqa: E402
     NO_MIR,  # noqa: E402
