+++ tests/regrid/regrid_mir/test_regrid_grib_message.py
@@ -8,7 +8,7 @@
 
 import pytest
 
-from earthkit.geo.regrid import regrid
+from earthkit.geo import regrid
 from earthkit.geo.utils.testing import (
     NO_EKD,  # noqa: E402
     NO_MIR,  # noqa: E402
