+++ tests/regrid/regrid_mir/test_regrid_numpy.py
@@ -9,7 +9,7 @@
 import numpy as np
 import pytest
 
-from earthkit.geo.regrid.array import regrid as regrid_array
+from earthkit.geo.grids.array import regrid as regrid_array
 from earthkit.geo.utils.testing import (
     NO_MIR,  # noqa: E402
     compare_global_ll_results,
