+++ tests/regrid/regrid_mir/test_regrid_xarray.py
@@ -10,7 +10,7 @@
 import numpy as np
 import pytest
 
-from earthkit.geo.regrid import regrid
+from earthkit.geo import regrid
 from earthkit.geo.utils.testing import (
     NO_EKD,  # noqa: E402
     NO_MIR,  # noqa: E402
